# [disasters] small page optimizations