--- a/server/config/disaster_dashboard/Earth.textproto
+++ b/server/config/disaster_dashboard/Earth.textproto
@@ -64,7 +64,7 @@ metadata {
         display_name: "Area"
         unit: "SquareKilometer"
         upper_limit: 1000
-        lower_limit: 25
+        lower_limit: 200
       }
       end_date_prop: "endDate"
       end_date_prop: "containmentDate"