--- a/server/config/disaster_dashboard/Earth.textproto
+++ b/server/config/disaster_dashboard/Earth.textproto
@@ -46,7 +46,7 @@ metadata {
         display_name: "Area"
         unit: "SquareKilometer"
         upper_limit: 1000
-        lower_limit: 200
+        lower_limit: 100
       }
     }
   }
@@ -64,7 +64,7 @@ metadata {
         display_name: "Area"
         unit: "SquareKilometer"
         upper_limit: 1000
-        lower_limit: 200
+        lower_limit: 100
       }
       end_date_prop: "endDate"
       end_date_prop: "containmentDate"