--- a/server/integration_tests/test_data/demo_feb2023/query_2/chart_config.json
+++ b/server/integration_tests/test_data/demo_feb2023/query_2/chart_config.json
@@ -45,7 +45,7 @@
           "color": "#f01212",
           "defaultSeverityFilter": {
             "displayName": "Area",
-            "lowerLimit": 25.0,
+            "lowerLimit": 100.0,
             "prop": "area",
             "unit": "SquareKilometer",
             "upperLimit": 1000.0