--- a/server/routes/api/disaster_api.py
+++ b/server/routes/api/disaster_api.py
@@ -21,6 +21,8 @@ from flask import current_app
 from flask import request
 from flask import Response
 
+import server.lib.util as lib_util
+from server.cache import cache
 import server.services.datacommons as dc
 
 # Define blueprint
@@ -197,6 +199,7 @@ def json_event_data():
 
 
 @bp.route('/event-data')
+@cache.cached(timeout=3600 * 24, query_string=True)  # Cache for one day.
 def event_data():
   """Gets the event data for a given eventType, date range, place, and
       filter information (filter prop, unit, lower limit, and upper limit).
@@ -264,4 +267,4 @@ def event_data():
             "provenanceInfo": provenance_info
         }
     }
-  return Response(json.dumps(result), 200, mimetype='application/json')
+  return lib_util.gzip_compress_response(result, is_json=True)