- add gzip compression to /geojson and /event-data
- add caching to /event-data
- increase lower limit of fire and flood severity

(third bar from top is floods, bottom bar is geojson. both for earth)

without compression:
![image](https://user-images.githubusercontent.com/6052978/218891272-87d31ee5-eb0b-467e-816e-f80120931fa6.png)

after compression:

compression level: 1 (flood - 2.85s waiting, wildfire - 2.09s, drought - 2.03s)
![image](https://user-images.githubusercontent.com/6052978/218892263-1cc60fa3-390b-45dd-a0a6-1cc7c2c56146.png)

compression level: 3 (flood - 2.74s waiting. wildfire - 2.10s, drought - 2.35s)
![image](https://user-images.githubusercontent.com/6052978/218892182-97743575-d0c8-4df0-b6c2-c232d4a6f61a.png)

caching without compression: download time 171.44ms
![image](https://user-images.githubusercontent.com/6052978/218891585-1691c7a6-36d6-4c65-9b47-39200194d461.png)

caching with compression (level 3): download time 121.96ms
![image](https://user-images.githubusercontent.com/6052978/218891677-e4ba09e8-990b-4732-b065-0cca29553de7.png)