@beets 2023-02-15
> this is cool! Just curious, are there no client side changes needed to read the compressed data?

there isn't! it's part of the http spec